This is a propossal for a basic [Dependabot](https://github.com/dependabot) configuration.
It scans the `.github/workflows/` directory for actions and will create PRs proposing updates if any of the actions have newer releases. It is configured to weekly scheduled runs.

Dependabot can be extended to support more, this focus of this PR is on GitHub Actions, where I have found Dependabot very useful in keeping my supply chain up to date.

More information is also available in the [GitHub Documentation](https://docs.github.com/en/code-security/supply-chain-security/managing-vulnerabilities-in-your-projects-dependencies/about-alerts-for-vulnerable-dependencies)