Update paper-advanced.md
Fixed the hotlinking.

I also just eliminated the pictures. The anchor was intentional because I thought it looked better, but there is really no reason to have the pictures for the video anyway. I also think it might be a little confusing to include them after all.
I also fixed the Wi-Fi spelling error.